requirements.txt
pymodbus
pyserial